+++ neoforge/src/main/kotlin/FallingTreesNeoForge.kt
@@ -12,6 +12,8 @@
 package dev.pandasystems.fallingtrees.neoforge
 
 import dev.pandasystems.fallingtrees.FallingTrees
+import dev.pandasystems.fallingtrees.compat.ModCompatibilities
+import dev.pandasystems.fallingtrees.neoforge.compat.TreeChopCompatImpl
 import net.neoforged.bus.api.IEventBus
 import net.neoforged.fml.common.Mod
 
@@ -19,5 +21,8 @@ import net.neoforged.fml.common.Mod
 class FallingTreesNeoForge(eventBus: IEventBus) {
 	init {
 		FallingTrees
+
+		if (ModCompatibilities.isTreeChopLoaded)
+			TreeChopCompatImpl.init(eventBus)
 	}
 }